egret/requirements.txt
ase
torch
mace-torch==0.3.12
